# perf: all: remove unnecessary allocations from strings.Builder.WriteString(fmt.Sprintf(...))