--- a/cosmovisor/args.go
+++ b/cosmovisor/args.go
@@ -379,7 +379,7 @@ func (cfg Config) DetailString() string {
 	var sb strings.Builder
 	sb.WriteString("Configurable Values:\n")
 	for _, kv := range configEntries {
-		sb.WriteString(fmt.Sprintf("  %s: %s\n", kv.name, kv.value))
+		fmt.Fprintf(&sb, "  %s: %s\n", kv.name, kv.value)
 	}
 	sb.WriteString("Derived Values:\n")
 	dnl := 0
@@ -390,7 +390,7 @@ func (cfg Config) DetailString() string {
 	}
 	dFmt := fmt.Sprintf("  %%%ds: %%s\n", dnl)
 	for _, kv := range derivedEntries {
-		sb.WriteString(fmt.Sprintf(dFmt, kv.name, kv.value))
+		fmt.Fprintf(&sb, dFmt, kv.name, kv.value)
 	}
 	return sb.String()
 }