--- a/x/distribution/types/proposal.go
+++ b/x/distribution/types/proposal.go
@@ -8,11 +8,11 @@ import (
 // String implements the Stringer interface.
 func (csp CommunityPoolSpendProposal) String() string {
 	var b strings.Builder
-	b.WriteString(fmt.Sprintf(`Community Pool Spend Proposal:
+	fmt.Fprintf(&b, `Community Pool Spend Proposal:
   Title:       %s
   Description: %s
   Recipient:   %s
   Amount:      %s
-`, csp.Title, csp.Description, csp.Recipient, csp.Amount))
+`, csp.Title, csp.Description, csp.Recipient, csp.Amount)
 	return b.String()
 }